Add files via upload
Today(7/2/25), I practiced the concept of file handling in Python. I worked on writing data to a file, opening the file in read mode, reading data from the file, appending data to the file, and copying the data to another file. This helped me understand how to efficiently manage file operations in Python.